added error channels to config

+++ core/admin.py
@@ -17,8 +17,8 @@ class Admin(commands.Cog):
 
     @commands.command()
     @has_permissions(administrator=True)
-    async def purge(self, ctx, limit: int=0):
-        await ctx.channel.purge(limit=limit+1)
+    async def purge(self, ctx, limit: int = 0):
+        await ctx.channel.purge(limit=limit + 1)
 
     """--------------------------------------------------------------------------------------------------------------------------"""
 
@@ -52,7 +52,8 @@ class Admin(commands.Cog):
             await ctx.send("channel is not a category", delete_after=5)
 
         for channel in category.channels:
-            await channel.delete()
+            if not channel.last_message_id:
+                await channel.delete()
         await category.delete()
 
     @commands.group(name="log_channel")
@@ -67,6 +68,25 @@ class Admin(commands.Cog):
             local_db.setdefault("log_channels", [])
             if ctx.channel.id not in local_db["log_channels"]:
                 local_db["log_channels"].append(ctx.channel.id)
+
+        with open("assets/local_db.json", "w", encoding="utf-8") as file:
+            json.dump(local_db, file)
+
+        await ctx.delete()
+        await ctx.send("Log Channel set successfully", delete_after=5)
+
+    @commands.group(name="error_channel")
+    @has_permissions(manage_channels=True)
+    async def error_channel(self, ctx):
+        pass
+
+    @error_channel.command(name="set")
+    async def error_channel_set(self, ctx):
+        with open("assets/local_db.json", "r", encoding="utf-8") as file:
+            local_db = json.load(file)
+            local_db.setdefault("error_channels", [])
+            if ctx.channel.id not in local_db["error_channels"]:
+                local_db["error_channels"].append(ctx.channel.id)
 
         with open("assets/local_db.json", "w", encoding="utf-8") as file:
             json.dump(local_db, file)
